doc-extract-filter-skill/requirements.txt
# 项目依赖

# 核心依赖
PyPDF2==3.0.1
python-docx==0.8.11
openpyxl==3.1.2

# CLI
click==8.1.3
